requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2026.1.4
charset-normalizer==3.4.4
fairscale==0.4.13
filelock==3.20.3
fsspec==2026.1.0
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
portalocker==3.2.0
pyrealsense2==2.56.5.9235
PyYAML==6.0.3
ram @ git+https://github.com/xinyu1205/recognize-anything.git@7cb804a8609e9f4b1a50b7f31436d2df40bb9481
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
scipy==1.17.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==80.9.0
sympy==1.14.0
timm==1.0.24
tokenizers==0.15.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.35.2
triton==3.5.1
typing_extensions==4.15.0
urllib3==2.6.3
